updates to main file
Removed the lives int, as the game does not require lives but only for players to score points (this was left over from Hortons instructions on how to build the pong clone). Created game over text and a win condition that requires one of the two players to score 7 total points. Upon scoring 7 points, a "game over" screen appears.